(Shafi) Bug fixed, all set
