+++ src/context/AuthContext.js
@@ -26,7 +26,6 @@ export const AuthContextProvider=({children})=>{
    useEffect(()=>{
       const unsubscribe=onAuthStateChanged(auth,(currentUser)=>{
          setUser(currentUser);
-         console.log(currentUser);
       })
       return ()=>{
          unsubscribe();
